// src/enums/history.enum.ts
export enum HistoryService {
  EARN = 'earn',
  PAYMENT = 'payment',
}

export enum PaymentHistory {
  PAYMENT = 'payment',
  WITHDRAW = 'withdraw',
  REFERRAL = 'referral',
  DONATE = 'donate',
}

export enum OfferStatus {
  ALL = 'all',
  PRE_APPROVED = 'pre_approved',
  TERMINATED = 'terminated',
  SCREEN_OUT = 'screen_out',
  COMPLETED = 'completed',
  PENDING = 'pending',
  REJECTED = 'rejected',
  PAID = 'paid',
}